remove redundant test file
